# openapi/ema.v1.s16a.json
{
  "components": {
    "schemas": {
...
      "ConfirmInput": {
        "properties": {
          "confirm": {
            "title": "Confirm",
            "type": "boolean"
          }
        },
        "required": [
          "confirm"
        ],
        "title": "ConfirmInput",
        "type": "object"
      },
...
      "DeleteResult": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "DeleteResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "client_slug": {
            "title": "Client Slug",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "type": {
            "enum": [
              "invoices",
              "piee",
              "audit",
              "reporting"
            ],
            "title": "Type",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "type",
          "client_slug",
          "year",
          "state"
        ],
        "title": "Job",
        "type": "object"
      },
...
      "NewJob": {
        "properties": {
          "client": {
            "title": "Client",
            "type": "string"
          },
          "type": {
            "enum": [
              "invoices",
              "piee",
              "audit",
              "reporting"
            ],
            "title": "Type",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "type",
          "client"
        ],
        "title": "NewJob",
        "type": "object"
      },
      "NewJobResult": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "NewJobResult",
        "type": "object"
      },
...
      "SectionState": {
        "properties": {
          "changed_input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed Input"
          },
          "fact_revisions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Revisions"
          },
          "fingerprint": {
            "items": {
              "type": "string"
            },
            "title": "Fingerprint",
            "type": "array"
          },
          "material_inputs": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Inputs"
          },
          "na_applicable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Na Applicable"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "section_id": {
            "title": "Section Id",
            "type": "string"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "missing",
              "ready",
              "drafted",
              "done",
              "later",
              "n/a"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "section_id",
          "status",
          "revision",
          "stale",
          "reason",
          "fingerprint",
          "fact_revisions",
          "material_inputs",
          "changed_input",
          "na_applicable"
        ],
        "title": "SectionState",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Ema local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "get": {
        "operationId": "jobs_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response Jobs Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "new_job_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewJob"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewJobResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_job_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "job_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/jobs/{job_id}/conflicts/{conflict_id}": {
      "post": {
        "operationId": "handler_jobs__job_id__conflicts__conflict_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "conflict_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "candidate_id": "candidate-exemplu",
                "on_revision": 1
              },
              "schema": {
                "properties": {
                  "candidate_id": {
                    "type": "string"
                  },
                  "on_revision": {
                    "type": "integer"
                  }
                },
                "required": [
                  "candidate_id",
                  "on_revision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "chosen": "candidate-exemplu",
                  "decision_id": "decision-exemplu",
                  "revision": 2
                },
                "schema": {
                  "properties": {
                    "chosen": {
                      "type": "string"
                    },
                    "decision_id": {
                      "type": "string"
                    },
                    "revision": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "decision_id",
                    "chosen",
                    "revision"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
...
    "/jobs/{job_id}/outputs": {
      "get": {
        "operationId": "handler_jobs__job_id__outputs_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "edited_externally": false,
                    "id": "output-exemplu",
                    "kind": "draft",
                    "version": 1
                  }
                ],
                "schema": {
                  "items": {
                    "properties": {
                      "edited_externally": {
                        "type": "boolean"
                      },
                      "id": {
                        "type": "string"
                      },
                      "kind": {
                        "type": "string"
                      },
                      "version": {
                        "type": "integer"
                      }
                    },
                    "required": [
                      "id",
                      "version",
                      "kind",
                      "edited_externally"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
...
    "/jobs/{job_id}/sections": {
      "get": {
        "operationId": "get_sections_jobs__job_id__sections_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SectionState"
                  },
                  "title": "Response Get Sections Jobs  Job Id  Sections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sections",
        "tags": [
          "audit"
        ]
      },
      "patch": {
        "operationId": "handler_jobs__job_id__sections_patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "on_revision": 1,
                "section_id": "ch1",
                "status": "ready"
              },
              "schema": {
                "properties": {
                  "on_revision": {
                    "type": "integer"
                  },
                  "section_id": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "section_id",
                  "status",
                  "on_revision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "ch1",
                    "revision": 2,
                    "status": "ready"
                  }
                ],
                "schema": {
                  "items": {
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "revision": {
                        "type": "integer"
                      },
                      "status": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "status",
                      "revision"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
...
    "/jobs/{job_id}/sections/{section_id}/na-proposal": {
      "post": {
        "operationId": "handler_jobs__job_id__sections__section_id__na_proposal_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "section_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "on_revision": 1,
                "reason": "fără date aplicabile"
              },
              "schema": {
                "properties": {
                  "on_revision": {
                    "type": "integer"
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "reason",
                  "on_revision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "proposed_status": "n/a",
                  "reason": "fără date aplicabile",
                  "requires_human": true,
                  "section_id": "ch5.electric"
                },
                "schema": {
                  "properties": {
                    "proposed_status": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    },
                    "requires_human": {
                      "type": "boolean"
                    },
                    "section_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "section_id",
                    "proposed_status",
                    "reason",
                    "requires_human"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
...
    "/jobs/{job_id}/stages/{stage}": {
      "post": {
        "operationId": "handler_jobs__job_id__stages__stage__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "stage",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "on_revision": 1
              },
              "schema": {
                "properties": {
                  "on_revision": {
                    "type": "integer"
                  }
                },
                "required": [
                  "on_revision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "done": 1,
                  "run_id": "run-exemplu",
                  "stage": "draft",
                  "state": "running",
                  "total": 7
                },
                "schema": {
                  "properties": {
                    "done": {
                      "type": "integer"
                    },
                    "run_id": {
                      "type": "string"
                    },
                    "stage": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "total": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "run_id",
                    "stage",
                    "state",
                    "done",
                    "total"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
...
  }
}